Specify dependencies according to semver
If the Cargo.toml doesn't specify an operator for crate versions, the
caret operator (^) is used
(https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html).
This means if you specify a version like "1.2.3", cargo interprets that
as "^1.2.3", which means it is free to update to any larger version 1.y.z,
according to the rules of semantic versioning
(https://github.com/steveklabnik/semver#requirements).

This is a sensible default, since it allows getting bugfix updates for
dependencies without having to update their versions every time.
However, the implicit caret requirement is not obvious. People might
assume that a version spec like "1.2.3" pins the dependency to exactly
this version.

To avoid this confusion, this commit truncates the dependency versions
to only the relevant parts, i.e. those that won't change under
semver-compatible updates.