Fall back to <CFName> for partition name
Some DMGs observed in the wild miss the partition <Name> tags in their
plist, which makes dmgwiz return empty names for them. However, in most
(all?) of these DMGs have <CFName> tags instead, which contain name
strings in the same format as the <Name> tags.

This commit adds a fallback to use <CFName> if <Name> is not present.