Avoid temporary buffers for decompression
Some DMG files contain very large chunks (in the order of multiple GB).
Decompressing those into a buffer means that dmgwiz can use an excessive
amount of memory or even crash. To fix this, this commit removes the
buffers and instead makes the decompression work on the input and output
`Read` and `Write` objects directly.

This currently doesn't work for LZFSE and ADC compression because their
implementations are currently lacking some support for buffer-less usage.