Revert to using buffers for lzfse decoding
The approach of implementing a streaming `Read`er for lzfse decoding
doesn't work unfortunately as it turns out that the lzfse algorithm
expects to be able to access the decoded bytes of previous chunks.

This commit therefore reverts back to using memory buffers for lzfse
decoding. Other compression algorithms are still handled in a streaming
way.