Fix broken library due to new plist 1.5 release.
The new plist 1.5 has some change in its xml dependencies, which causes this library to break. You can simulate that my running `cargo update && cargo test` on the main branch.

The proper fix would be to see if we can adapt the code to the new xml dependency version,
however just restricting the plist version to last know good one is an easy fix for now.